requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
backports.tarfile==1.2.0; python_version < '3.12'
cachetools==5.3.3
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
deprecated==1.2.13
docutils==0.21.2; python_version >= '3.9'
flask-cors==4.0.1
flask-limiter==3.1.0
flask==2.2.2
greenlet==2.0.2
idna==3.4
importlib-metadata==7.2.0; python_version >= '3.8'
itsdangerous==2.1.2
jaraco.classes==3.4.0; python_version >= '3.8'
jaraco.context==5.3.0; python_version >= '3.8'
jaraco.functools==4.0.1; python_version >= '3.8'
jinja2==3.1.2
keyring==25.2.1; python_version >= '3.8'
limits==3.2.0
markupsafe==2.1.2
more-itertools==10.3.0; python_version >= '3.8'
nh3==0.2.17
numpy==1.24.1
ordered-set==4.1.0
packaging==23.2; python_version >= '3.7'
pandas==1.5.3
patsy==0.5.6
pkginfo==1.11.1; python_version >= '3.8'
psycopg2==2.9.5
pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7.1
pywin32-ctypes==0.2.2; sys_platform == 'win32'
readme-renderer==43.0; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.28.2
rfc3986==2.0.0; python_version >= '3.7'
rich==12.6.0
scipy==1.13.1
setuptools==70.1.0; python_version >= '3.8'
six==1.16.0
sqlalchemy==2.0.1
statsmodels==0.14.2
texas-hold-em-utils==0.0.4
twine==5.1.0
typing-extensions==4.4.0
urllib3==1.26.14
waitress==3.0.0
werkzeug==2.2.2
wrapt==1.14.1
zipp==3.19.2; python_version >= '3.8'
